issue/#212 move components from my page to global
