+++ components/InactiveRallies/InactiveRallyCard/CompletedLabel/index.tsx
